fix some merge errors
